applied the patch from the 2.9.0 branch, and fixed service number
git-svn-id: http://svn.xp-dev.com/svn/wxextension@846 f22100f3-aa73-48fe-a4fb-fd497bb32605

+++ trunk/syncsocketserver/src/server.cpp
@@ -796,8 +796,8 @@ const wxString MyFrame::SocketDetails(const wxSocketBase* sock) const
 
   value <<
     _("socket: ") <<
-    local_addr.IPAddress() << "." << local_addr.Service() << ", " <<
-    peer_addr.IPAddress() << "." << peer_addr.Service();
+    local_addr.IPAddress() << "." << (int)local_addr.Service() << ", " <<
+    peer_addr.IPAddress() << "." << (int)peer_addr.Service();
 
   return value;
 }
